requirements.txt
fastapi
uvicorn
docker
pydantic
python-dotenv
requests
